Move fixtures out of tests subfolder
