+++ ethereum_history_api/oracles/fixtures/mockClientData.json
