+++ ethereum_history_api/oracles/src/noir/oracles/accountOracle.test.ts
@@ -3,14 +3,14 @@ import { encodeAccount, encodeStateProof, parseNoirGetAccountArguments } from '.
 import { type GetProofReturnType } from 'viem';
 import { readFile } from 'fs/promises';
 import { parse } from '../../util/json-bigint.js';
-import account from '../../../test/fixtures/account.json';
-import stateProof from '../../../test/fixtures/stateProof.json';
+import account from '../../../fixtures/account.json';
+import stateProof from '../../../fixtures/stateProof.json';
 import { serializeAccount, serializeStateProof } from '../../../test/helpers.js';
 import { ADDRESS } from '../../ethereum/recordingClient.test.js';
 
 describe('encodeAccount', async () => {
   it('encode account', async () => {
-    const proof: GetProofReturnType = parse(await readFile('./test/fixtures/eth_getProof_response.json', 'utf-8'));
+    const proof: GetProofReturnType = parse(await readFile('./fixtures/eth_getProof_response.json', 'utf-8'));
 
     expect(encodeAccount(proof)).toStrictEqual(serializeAccount(account));
     expect(encodeStateProof(proof)).toStrictEqual(serializeStateProof(stateProof));
