+++ ethereum_history_api/oracles/src/script/encodeHeader.ts
@@ -1,4 +1,4 @@
-import { blockHeaders } from '../../test/fixtures/blockHeader.json';
+import { blockHeaders } from '../../fixtures/blockHeader.json';
 import { assert } from '../util/assert.js';
 import { type BlockHeader } from '../ethereum/blockHeader.js';
 import { encodeBlockHeaderPartial } from '../noir/oracles/headerOracle.js';
