+++ ethereum_history_api/oracles/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.common.json",
-  "include": ["src/**/*", "test/fixtures/*.json", "test/helpers.ts", "package.json"],
+  "include": ["src/**/*", "fixtures/*.json", "test/helpers.ts", "package.json"],
   "compilerOptions": {
     "rootDirs": ["src", "test"],
     "outDir": "dist"
